viajarfazbem/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-daterange-filter==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
filelock==3.0.12
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
openapi-codec==1.3.2
pbr==5.6.0
platformdirs==2.1.0
psycopg2==2.9.1
pycparser==2.20
pycryptodome==3.10.1
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.26.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
stevedore==3.3.0
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.6.0
virtualenv-clone==0.5.5
virtualenvwrapper==4.8.4
Werkzeug==1.0.1
whitenoise==5.3.0
zipp==3.4.1
